refactor: 슈퍼 DTO 제거
