+++ backend/src/main/java/zipgo/auth/presentation/dto/AuthResponse.java
@@ -4,8 +4,6 @@ import java.util.List;
 import zipgo.member.domain.Member;
 import zipgo.pet.domain.Pet;
 import zipgo.pet.presentation.dto.response.PetResponse;
-import zipgo.pet.presentation.dto.response.PetResponses;
-import zipgo.review.dto.response.GetReviewResponse;
 
 public record AuthResponse(
         String name,
