--- backend/src/main/java/zipgo/review/dto/response/RatingSummaryElement.java
@@ -1,10 +1,0 @@
-package zipgo.review.dto.response;
-
-import java.util.List;
-
-public record RatingSummaryElement(
-        double average,
-        List<SummaryElement> rating
-) {
-
-}
