--- backend/src/main/java/zipgo/review/dto/response/SummaryElement.java
@@ -1,8 +1,0 @@
-package zipgo.review.dto.response;
-
-public record SummaryElement(
-        String name,
-        int percentage
-) {
-
-}
