+++ backend/src/main/java/zipgo/review/presentation/dto/request/CreateReviewRequest.java
@@ -1,4 +1,4 @@
-package zipgo.review.dto.request;
+package zipgo.review.presentation.dto.request;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
