+++ backend/src/main/java/zipgo/review/presentation/dto/request/GetReviewsRequest.java
@@ -1,4 +1,4 @@
-package zipgo.review.dto.request;
+package zipgo.review.presentation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
