+++ backend/src/main/java/zipgo/review/presentation/dto/response/type/RatingInfoResponse.java
@@ -1,0 +1,8 @@
+package zipgo.review.presentation.dto.response.type;
+
+public record RatingInfoResponse(
+        String name,
+        int percentage
+) {
+
+}
