+++ backend/src/main/java/zipgo/review/presentation/dto/response/type/RatingSummaryResponse.java
@@ -1,0 +1,10 @@
+package zipgo.review.presentation.dto.response.type;
+
+import java.util.List;
+
+public record RatingSummaryResponse(
+        double average,
+        List<RatingInfoResponse> rating
+) {
+
+}
