+++ backend/src/main/java/zipgo/review/presentation/dto/response/type/StoolConditionResponse.java
@@ -1,0 +1,14 @@
+package zipgo.review.presentation.dto.response.type;
+
+import zipgo.review.domain.type.StoolCondition;
+
+public record StoolConditionResponse(
+        String name,
+        int percentage
+) {
+
+    public static StoolConditionResponse from(StoolCondition stoolCondition, int percentage) {
+        return new StoolConditionResponse(stoolCondition.getDescription(), percentage);
+    }
+
+}
