+++ backend/src/main/java/zipgo/review/presentation/dto/response/type/TastePreferenceResponse.java
@@ -1,0 +1,14 @@
+package zipgo.review.presentation.dto.response.type;
+
+import zipgo.review.domain.type.TastePreference;
+
+public record TastePreferenceResponse(
+        String name,
+        int percentage
+) {
+
+    public static TastePreferenceResponse of(TastePreference tastePreference, int percentage) {
+        return new TastePreferenceResponse(tastePreference.getDescription(), percentage);
+    }
+
+}
