+++ backend/src/test/java/zipgo/auth/presentation/AuthControllerMockArgumentResolverTest.java
@@ -37,8 +37,8 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWit
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static zipgo.member.domain.fixture.MemberFixture.식별자_있는_멤버;
-import static zipgo.pet.domain.fixture.BreedsFixture.*;
-import static zipgo.pet.domain.fixture.PetSizeFixture.*;
+import static zipgo.pet.domain.fixture.BreedsFixture.견종;
+import static zipgo.pet.domain.fixture.PetSizeFixture.대형견;
 
 
 @SuppressWarnings("NonAsciiCharacters")
