+++ backend/src/test/java/zipgo/review/fixture/ReviewFixture.java
@@ -6,8 +6,8 @@ import zipgo.pet.domain.Pet;
 import zipgo.petfood.domain.PetFood;
 import zipgo.review.domain.Review;
 import zipgo.review.domain.type.ReviewPetInfo;
-import zipgo.review.dto.request.CreateReviewRequest;
-import zipgo.review.dto.request.UpdateReviewRequest;
+import zipgo.review.presentation.dto.request.CreateReviewRequest;
+import zipgo.review.presentation.dto.request.UpdateReviewRequest;
 
 import static zipgo.review.domain.type.StoolCondition.DIARRHEA;
 import static zipgo.review.domain.type.StoolCondition.SOFT_MOIST;
